ec2_generic_policy/requirements.txt
aws-cdk.aws-ec2
aws-cdk.aws-ssm
aws-cdk.aws_iam
aws-cdk.aws-codepipeline
aws-cdk.aws-codepipeline_actions
aws-cdk.aws_codecommit
aws-cdk.pipelines